added user logon logoff and automatic guest signin

+++ app.py
@@ -34,9 +34,15 @@ def about():
 def reviews():
     tmdb_id = request.args.get('tmdb_id')
     reviews_exist = mongo.db.reviews.find_one({"tmdb_id" : tmdb_id})
+    if 'user' in session:
+        profile_id = session['user']
+    else:
+        profile_id = 'Guest'
+
     if reviews_exist:
         return render_template("reviews.html", 
-            tmdb_id=tmdb_id, 
+            tmdb_id=tmdb_id,
+            profile_id = profile_id, 
             movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
             reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))
             
@@ -44,6 +50,7 @@ def reviews():
         flash('No reviews have been added, be the first.', 'warning')
         return render_template('addreview.html', 
             tmdb_id=tmdb_id,
+            profile_id = profile_id,
             movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
             categories=mongo.db.categories.find(),
             ratings=mongo.db.ratings.find())
@@ -51,16 +58,22 @@ def reviews():
 @app.route("/addreview/<tmdb_id>")
 def addreview(tmdb_id):
     if 'user' in session:
-        flash('Logged in successfully as : ' + session['user'], 'success')
         return render_template('addreview.html', 
             tmdb_id=tmdb_id,
+            profile_id = session['user'],
             movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
             categories=mongo.db.categories.find(),
             ratings=mongo.db.ratings.find())
     else:
-        flash('Login or Sign up to add a review!', 'warning')
-        return render_template('login.html',
-                            first_movie=mongo.db.movies.find_one())
+        flash('You are using the Guest profile, Login/Register to prevent others editing/deleting your reviews', 'warning')
+        session['user'] = 'Guest'
+        return render_template('addreview.html', 
+            tmdb_id=tmdb_id,
+            profile_id = session['user'],
+            movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
+            categories=mongo.db.categories.find(),
+            ratings=mongo.db.ratings.find())
+ 
 
 
 @app.route("/insertreview", methods=["POST"])
@@ -94,14 +107,29 @@ def insertreview():
 
 @app.route('/editreview/<review_id>/<tmdb_id>')
 def editreview(review_id,tmdb_id):
-    the_review =  mongo.db.reviews.find_one({"_id": ObjectId(review_id)})
-    all_categories =  mongo.db.categories.find()
-    return render_template('editreview.html',
-                           review=the_review,
-                           tmdb_id=tmdb_id,
-                           movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
-                           categories=all_categories,
-                           ratings=mongo.db.ratings.find())
+    if 'user' in session:
+        the_review =  mongo.db.reviews.find_one({"_id": ObjectId(review_id)})
+        if session['user'] == the_review['username']:
+            all_categories =  mongo.db.categories.find()
+            return render_template('editreview.html',
+                            review=the_review,
+                            tmdb_id=tmdb_id,
+                            movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
+                            categories=all_categories,
+                            ratings=mongo.db.ratings.find())
+        else:
+            flash('You are logged on as ' + session['user'] + ' and can only EDIT those reviews!', 'warning')
+            return render_template("reviews.html", 
+            tmdb_id=tmdb_id, 
+            movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
+            reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))
+    else:
+        session['user'] = 'Guest'
+        flash('You are logged on as ' + session['user'] + ' and can only EDIT those reviews!', 'warning')
+        return render_template("reviews.html", 
+        tmdb_id=tmdb_id, 
+        movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
+        reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))
 
 @app.route('/updatereview/<review_id>/<tmdb_id>', methods=["POST"])
 def updatereview(review_id,tmdb_id):
@@ -118,29 +146,46 @@ def updatereview(review_id,tmdb_id):
     flash('Review updated successfully', 'success')
     return render_template("reviews.html", tmdb_id=tmdb_id,
                            movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
-                           reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }))
+                           reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))
 
 @app.route('/deletereview/<review_id>/<tmdb_id>', methods=["POST"])
 def deletereview(review_id,tmdb_id):
-    mongo.db.reviews.delete_one({'_id': ObjectId(review_id)})
-
-    mongo.db.movies.find_one_and_update(
-        {'tmdb_id': tmdb_id},
-        {'$inc': {'review_count': -1}}
-    )
-    reviews_exist = mongo.db.reviews.find_one({"tmdb_id" : tmdb_id})
-    if reviews_exist:
-        flash('Review deleted successfully!', 'success')
-        return render_template("reviews.html", 
+    if 'user' in session:
+        the_review =  mongo.db.reviews.find_one({"_id": ObjectId(review_id)})
+        if session['user'] == the_review['username']:
+            mongo.db.reviews.delete_one({'_id': ObjectId(review_id)})
+            mongo.db.movies.find_one_and_update(
+                {'tmdb_id': tmdb_id},
+                {'$inc': {'review_count': -1}}
+            )
+            reviews_exist = mongo.db.reviews.find_one({"tmdb_id" : tmdb_id})
+            if reviews_exist:
+                flash('Review deleted successfully!', 'success')
+                return render_template("reviews.html", 
+                    tmdb_id=tmdb_id, 
+                    movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
+                    reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))
+            else:
+                flash('The last review was deleted', 'success')
+                mongo.db.movies.remove({'tmdb_id': tmdb_id})
+                return render_template('home.html', 
+                    first_movie=mongo.db.movies.find_one(), 
+                    movies=mongo.db.movies.find().sort("last_updated", -1))
+        else:
+            flash('You are logged on as ' + session['user'] + ' and can only delete those reviews!', 'warning')
+            return render_template("reviews.html", 
             tmdb_id=tmdb_id, 
             movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
             reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))
     else:
-        flash('The last review was deleted', 'success')
-        mongo.db.movies.remove({'tmdb_id': tmdb_id})
-        return render_template('home.html', 
-            first_movie=mongo.db.movies.find_one(), 
-            movies=mongo.db.movies.find().sort("last_updated", -1))
+        session['user'] = 'Guest'
+        flash('You are logged on as ' + session['user'] + ' and can only DELETE those reviews!', 'warning')
+        return render_template("reviews.html", 
+        tmdb_id=tmdb_id, 
+        movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
+        reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))
+
+
 
     
     
@@ -150,6 +195,12 @@ def insertmovie():
     movie_url = request.form.get('form_poster_url')
     movie_in_collection = mongo.db.movies.find_one({"tmdb_id" : tmdb_id})
     reviews_exist = mongo.db.reviews.find_one({"tmdb_id" : tmdb_id})
+    if 'user' in session:
+        profile_id = session['user']
+    else:
+        session['user'] = 'Guest'
+        profile_id = 'Guest'
+
     if movie_in_collection:
         if reviews_exist:
             return render_template("reviews.html", 
@@ -157,9 +208,10 @@ def insertmovie():
                            movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
                            reviews=mongo.db.reviews.find( { 'tmdb_id': tmdb_id }).sort("review_date", -1))    
         else:
-            flash('No reviews have been added, be the first.', 'warning')
+            flash('You are logged on as ' + profile_id + ' there are no reviews, be the first.', 'warning')
             return render_template('addreview.html', 
                             tmdb_id=tmdb_id,
+                            profile_id=profile_id,
                             movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
                             categories=mongo.db.categories.find(),
                             ratings=mongo.db.ratings.find())        
@@ -175,9 +227,10 @@ def insertmovie():
                 'review_count':0
                 }
         movies.insert_one(post)
-        flash('No reviews have been added, be the first.', 'warning')
+        flash('You logged on as ' + profile_id + 'there are no reviews, be the first.', 'warning')
         return render_template('addreview.html', 
                             tmdb_id=tmdb_id,
+                            profile_id= profile_id,
                             movie=mongo.db.movies.find_one({"tmdb_id" : tmdb_id}),
                             categories=mongo.db.categories.find(),
                             ratings=mongo.db.ratings.find())  
@@ -207,7 +260,9 @@ def search():
 
 @app.route("/register")
 def register():
-   return render_template('register.html',
+    if 'user' in session:
+        session.pop('user')
+    return render_template('register.html',
                             first_movie=mongo.db.movies.find_one())
 
 @app.route('/registercheck', methods=['GET', 'POST'])
@@ -241,7 +296,9 @@ def registercheck():
 
 @app.route("/login")
 def login():
-   return render_template('login.html',
+    if 'user' in session:
+        session.pop('user')
+    return render_template('login.html',
                             first_movie=mongo.db.movies.find_one())
 
 @app.route('/logincheck', methods=['GET', 'POST'])
